# Improve Status performance with many ignored files