--- a/utils/merkletrie/difftree.go
+++ b/utils/merkletrie/difftree.go
@@ -252,6 +252,7 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/go-git/go-git/v5/plumbing/format/gitignore"
 	"github.com/go-git/go-git/v5/utils/merkletrie/noder"
 )
 
@@ -297,8 +298,14 @@ func DiffTreeContext(ctx context.Context, fromTree, toTree noder.Noder,
 		case noMoreNoders:
 			return ret, nil
 		case onlyFromRemains:
-			if err = ret.AddRecursiveDelete(from); err != nil {
-				return nil, err
+			if node, ok := ignoredNode(toTree, from); ok {
+				if err = diffNodesSameName(&ret, ii, ii.from.current, node); err != nil {
+					return nil, err
+				}
+			} else {
+				if err = ret.AddRecursiveDelete(from); err != nil {
+					return nil, err
+				}
 			}
 			if err = ii.nextFrom(); err != nil {
 				return nil, err
@@ -353,8 +360,10 @@ func diffNodes(changes *Changes, ii *doubleIter) error {
 	// compare their full paths as strings
 	switch from.Compare(to) {
 	case -1:
-		if err = changes.AddRecursiveDelete(from); err != nil {
-			return err
+		if ok := isIgnoredNode(to[0], from); !ok {
+			if err = changes.AddRecursiveDelete(from); err != nil {
+				return err
+			}
 		}
 		if err = ii.nextFrom(); err != nil {
 			return err
@@ -367,7 +376,7 @@ func diffNodes(changes *Changes, ii *doubleIter) error {
 			return err
 		}
 	default:
-		if err := diffNodesSameName(changes, ii); err != nil {
+		if err := diffNodesSameName(changes, ii, ii.from.current, ii.to.current); err != nil {
 			return err
 		}
 	}
@@ -375,11 +384,8 @@ func diffNodes(changes *Changes, ii *doubleIter) error {
 	return nil
 }
 
-func diffNodesSameName(changes *Changes, ii *doubleIter) error {
-	from := ii.from.current
-	to := ii.to.current
-
-	status, err := ii.compare()
+func diffNodesSameName(changes *Changes, ii *doubleIter, from, to noder.Path) error {
+	status, err := ii.compareNoders(from, to)
 	if err != nil {
 		return err
 	}
@@ -450,4 +456,19 @@ func diffDirs(changes *Changes, ii *doubleIter) error {
 	}
 
 	return nil
+}
+
+func isIgnoredNode(tree noder.Noder, path noder.Path) bool {
+	in, ok := tree.(*gitignore.MatchNoder)
+
+	return ok && in.PathIgnored(path)
+}
+
+func ignoredNode(tree noder.Noder, path noder.Path) (noder.Path, bool) {
+	in, ok := tree.(*gitignore.MatchNoder)
+	if !ok || !in.PathIgnored(path) {
+		return nil, false
+	}
+
+	return in.FindPath(path)
 }